[move-package] Don't update tag based git dependencies (#662)
Tag based git dependencies can't really be updated easily after
the initial download, so these will not be updated.  Tags are
usually used as releases, and in general shouldn't change
after the fact.  This becomes tricky in general in comparison
to a branch based flow.  Additionally, added an updating message
and cleaned up some duplicate outputs.