[move-model] two ways to test is_intrinsic
A function can be marked as intrinsic in two ways:

Method 1:

```
spec foo {
    pragma intrinsic;
}
```

Method 2:

```
spec Table {
    pragma intrinsic = map,
        map_foo = foo;
}
```

In both ways, `foo` should be considered as intrinsic while the current
approach only treats the first one as intrinsic and forces everything
declared as a intrinsic function of some struct to be re-marked as
intrinsic.

This PR changes this behavior. `foo` will be recognized as intrinsic
regardless of how the intrinsic property is declared.