[move-prover] allow annotations to mark whether a fixedpoint is reached
This commit allows the `Annotations` struct which resides in
`FunctionData` to mark whether an annotation has reached a fixedpoint
status in the presence of recursive calls.

The core of change is this:

```diff
pub struct Annotations {
-    map: BTreeMap<TypeId, Data>,
+    map: BTreeMap<TypeId, (Data, bool)>,
}
```

Currently all annotations, when set, are marked as already reached
fixedpoint status. This is consistent with the current assumption of
no-recursive calls.

Subsequent commits will relax this unsound assumption by examining the
annotation structure produced by different passes in the transformation
pipeline.

This commit depends on PR #698