[move-prover] convert borrow analysis with interprocedural absint
The key idea is to define a join function for the `BorrowAnnotation`
struct. The interprecedual absint for borrow analysis is following:

For all functions in a recursive call group (an SCC in the call graph):

- initially all function have a `BorrowAnnotation::default()` info
- for each function visited, produce a summary of `BorrowAnnotation`
  and *join* it with the existing summary (or the `default()` one.)
  - if `JoinResult::Unchanged` --> mark this annotation as reaching a
    fixedpoint state
  - if `JoinResult::Changed` --> mark this annotation as not yet
    reaching a fixedpoint.

The borrow analysis will stop when the borrow annotation summary for
all functions in the SCC are marked as reached fixedpoint.