[4/x][move-package/lock] Fix error message from parse_package_manifest
Fix the order in which the directory and `Move.toml` are joined
together (should be directory and then `Move.toml`, not the other way
around as it previously was), and don't drop the context of the
underlying error.

Also restructuring the function slightly to be less deeply nested, by
using `?`.

Test Plan:

```
move/language/tools/move-package$ cargo nextest
```