[tools] print imports, fully qualified function names in disassembler (#757)
Previously, the disassembler did not print dependency information or use fully qualified function names. That makes it difficult to see what code a module is using/which functions it's calling, and (in the case of name conflicts) can create ambiguity.

- Add a module alias map to the disassembler state that disambiguates between modules with the same name
- Use the alias map to print dep info, using source langauge syntax. Resolve name conflicts using `as` (again, like the source langguage).
- Also print fully qualified function names. Internal function calls have no module name prefix or `::`, as in the source language.

See added test for details on what the printed output looks like + how conflicts are resolved.