[borrow v2] Move borrow semantics v2 (type checker only)
This PR is a first attempt towards implementing borrow semantics v2 for Move. In a nutshell, we implement the capability that storage references (als call _static_ references) can (a) travel upwards the call stack and (b) leave module boundaries.

For background, it is currently not possible in Move to write a function `fun my_borrow(..): &R { borrow_global<R>(...) }`. Instead the result of `borrow_global` and `borrow_global_mut` can only be passed downwards the stack. This is a strong restriction regards the ability to factorize code. For example, one cannot write any helper functions to access global storage easily.

In this PR we avoid the complexity of full Rust style lifetimes and instead evolve the existing `acquires` concept of Move. The `acquires R` annotation on functions is already needed to declare that a function accesses global storage. We now allow functions which `acquire R` to potentially return `&R` or `&mut R`:

```
fun my_borrow(a: address): &R acquires R {
  borrow_global<R>(a)
}
```

On caller side of `my_borrow`, we can see that the function returns a reference to resource storage and has no incoming reference, henceforth a borrow edge is derived which notes that R is in fact derived from storage. This mechanism propagates upwards the call stack.

We also allow storage references to travel outside of module boundaries. This extends the `acquires` mechanism from local structs to imported structs. Currently, acquires for calls to other modules does not need to be considered because references to resources managed by thiose modules cannot exist. This is kept as the default behavior, but a user can mark a struct with the attribute `#[open_struct]` to indicate that references to it in storage can leave the module, in which cause `acquires` need to be used for external calls.

In addition to the `#[open_struct]` attribute, this PR introduces a new attribute `#[no_borrow(param1, ..., paramn)]` which allows to exclude an input borrow for implicitly creating a borrow edge. Without this, we wouldn't be able to describe for instance functions as the following:

```
fun reader_ref<T>(r: &ReaderRef<T>): &T { borrow_global<T>(r.addr) }
```

Here the parameter `r` is _not_ a borrowing root for `&T`, instead we want to express that `&T` is from storage. Without the `no_borrow` attribute this would not be possible.