[move-lang] Inlined functions (formerly macros): parsing and type checking (#774)
```
    public inline fun foreach<T>(v: &vector<T>, action: |&T|) {
        let i = 0;
        while (i < Vector::length(v)) {
            action(Vector::borrow(v, i));
            i = i + 1;
        }
    }
```

This PR implements parsing and type checking of this style of functions up to the typing AST. Macro expansion is not yet implemented and intended to be added in a followup.

The biggest change is enabling function types, lambda expressions, and calls of function values for naming and typing. Those had been supported before for the spec language, but weren't supported in move-lang beyond expansion. In the design, we represent function types by a new BuiltinTypeName_::Fun which allows to minimize changes to the code, as no new type variant needs to be introduced, and existing functionality for inference et. al can be adopted with minimal changes.

After this PR, to complete this style of functions, we still need to:

- Implement actual inlining. This is intended to happen in a new phase after typing.
- Implement techniques (in the prover) to refer to specifications of passed lambda arguments to inlined functions. (See PR #9978 for what we want to do here.)