[move-prover] data invariant assume and assertion for `&mut` params
For functions with `&mut` params `s: S`, the data invariant for `S` will
be assumed at function entry (via `Wellformedness` - this is already
done) and re-checked on any write-back to this reference --- this is
newly added in this commit.

This PR is the opposite of PR #783 which does the opposite, i.e., drop
the assumption at the function entry. PR #783 is proposed following the
following guideline:

> Data invariants are only checked when the `&mut` borrow ends. During
> the lifetime of the `&mut` reference, the invariant can be temporarily
> violated and hence, assuming data invariants there can lead to
> inconsistency issues.

In theory, both PR's fix are better than the current practice (which
is essentially assuming without assertion).

However, as the test result in PR #783 shows, the `Options.move` in
`stdlib` heavily relies on the entrypoint assumption that a `&mut
Option` has either 0 or 1 entry in its vector. Therefore, in this
commit, we tighten the conditions where we assert data invariant.

In summary, after this PR, data invariants will be checked:

- on write-backs to local root, arised from `let r = &mut s` where `s`
  is a local variable
- on write-backs to global root, arised from `let r =
  borrow_global_mut<S>(addr)`
- on write-backs to `&mut` parameter.

```move
struct S {
    f: u8
}
spec S {
    invariant f != 0;
}

fun foo(s: &mut S) {
    // spec: assume s.f != 0;

    let r = &mut s.f;
    r = 0;

    // spec: write-back[s.f](r);
    // spec: assert s.f != 0;
}
```