Cherry picks TypeTag bounds from old aptos branch (#791)
* [types] make type tag nesting limited

There's limited value and a lot of wasted resources by leaving this
limited to serde's limits. Recursion is dangerous.

* [types] limit nesting of types via state (#532)

Cherry picked from #529.

This is an alternative version of what #528 and the previous for limitation of nesting provides, having a significant lower footprint.

This approach installs custom serializers at the recursion points of the `TypeTag` enum. Those serializer use a thread local to track the nesting of the type tag, both for serialization and deserialization.

The usage of a thread local is safe and a common technique to provide extra state for implementations which miss it (e.g. serde). A given thread has a single execution stack and this will always count the depth counter consistently up and down. Because Rust has no exceptions are other exit points, this cannot be corrupted. Possibly, thread locals may cause inefficiency. However, we expect TypeTags to be small (a few constructors, limited nestings).

The technique used for type tags can be generalized for any other recursive type if needed. Eventually, we hope serde will support per-type nesting, and possibly we should talk with the author if we can add it ourselves.

This uses and passes the tests from #528.

* [types] limit parsing of nested TypeTag

There's limited value and a lot of wasted resources by leaving this
limited to serde's limits. Recursion is dangerous.

Co-authored-by: David Wolinsky <isaac.wolinsky@gmail.com>