[inlining] Expansion of inlined functions (#795)
* [inlining] Expansion of inlined functions

This PR implements the core logic for expanding inlined functions into the AST. Expansion is happening as a new phase after typing and before hlir lowering.  The inlining phase is a function `f(&mut naming::ast::Program)`. It will inline functions as they are encountered. When a function with lambda arguments is inlined, those lambdas are further expanded, using basically the same algorithm as for conventional inlining.

There are a few tests added which cover basic behavior. More tests have to be added as the feature evolves, and more type checks have to be added. Specifically, currently inlining can create code which does not pass the bytecode verifier because 'private' calls are inserted at inlining context. This is not captured by the Move compiler, but the bytecode verifier will flag this. (With a terrible DevX.)

* fixup: remove debug comment