[inlining] Fixing bugs and adding first combinators to stdlib
This adds `map` and `filter` to `std::option` and make them work e2e, including unit tests in `option_tests`. At this point, we do have inlined functions working across modules and packages.

The PR also explores some conceptual questions around parameter expansion. Right now, we stick to the original simplistic viewpoint that all parameters are replaced directly by the according expression. This effectively implements 'call by expression'. It would be the responsibility (and flexibility!) of the function implementer to use `let` and similar measures to avoid duplicate evaluation of inlined parameters. This also creates a restriction: if somewhere in the code a parameter `p` is used in l-value position than also the passed parameter must be an lvalue. There is a new error message if this is not the case in `parameter_borrow.move`. An implementer can work around this restriction by avoiding `&param`, as done in the new combinators for options.