[move-prover] model type_name::get in stdlib
The majority of the Prover work was firstly introduced when Aptos
introduced the `type_info` Move module. Now that the stdlib has a
similar module `type_name`, we should bridge the effort here.

This PR is still a bit rough as it is using hardcoded module address /
function names to identify the type-reflection schemes. We should come
up with more robust / extensible approach later.