[RFC][2/x][move-package/lock] Extend handling of Dependency subdir (#745)
Proposing a more general handling of the `subdir` field for
dependencies, which boils down to:

- Explicitly disallowing the field on `local` dependencies (this will
produce an error, where previously it would be silently ignored).
- Enabling its use with custom dependencies.

The second change is not likely to be used directly in
`Move.toml` (although it could be), but instead is intended to be used
with lock files (which are designed to serialize/deserialize
dependencies in a format that can be parsed by the existing
`manifest_parser::parse_dependency`), in the context of dependency
re-rooting, which already occurs implicitly for `git` dependencies, when:

- `a`, depends on `b`, a git dependency, with subdir `s`, and
- `b`, depends on `c`, a local dependency, with relative path `r`.

From `a`'s perspective, `c` is a git dependency with `subdir`
`s.push(r)`, which should still be a path within `b`'s git repo (if it
is not, resolution will fail, but with an internal error).

When introducing a lock file, this operation becomes explicit, i.e. in
the lock file for `a`, `c` will appear as a git dependency with the
aforementioned subdir, which is referred to as "re-rooting" `c`, and
the error case of `c`'s path falling outside of `b`'s will become an
explicit error.

This raises a question for custom dependencies -- can they have local
transitive dependencies?  If so, they need a `subdir` field to be
written to the lock file, which this commit proposes.  If not, that
case can be treated as an error, but this seems overly restrictive.

## Test Plan

New unit test:

```
move-package$ cargo nextest
```