[6/x][move-package/lock] LockFile and DependencyGraph (#760)
Introduce `DependencyGraph` as the in-memory representation of a
`BuildConfig`-agnostic transitive dependency graph, and `LockFile`
which is a handle for the lock file that this graph can be serialized
into and deserialized out of.

The lock file is created at the start of a build as a temporary file
in a directory within `build`, and can optionally be committed at a
path supplied in `BuildConfig`.  If the explicit `commit` API is not
called, its contents will be discarded.  This means in the case of a
crash or abort, the lock file will not be left in an inconsistent
state, although if this happens and `LockFile`'s `drop` handler is not
called, a temporary file may leak in the `build` directory. (This is
usually cleaned up by a later "clean" operation in a subsequent run of
the compiler).

`lock_file::schema` comes with `serde`-compatible times for
deserializing the lock file, reading its version to determine
compatibility, and extracting the serialized dependencies from within
it (to be further parsed into `parsed_manifest::Dependency`).

This commit adds `DependencyGraph` and lock file generation as part of
package resolution, to write snapshot tests for it, but it does not
replace similar logic in `resolution_graph` (for exploring a package's
set of transitive dependencies), to come in a later commit.

Because `DependencyGraph` is `BuildConfig`-agnostic, this commit
introduces a behaviour change in that a package whose transitive
dependency graph is inconsistent in dev-mode (contains conflicting
dependencies) will not compile in normal mode.

## Test Plan

New snapshot tests for lock files, and existing CLI and
move-package/resolution tests:

```
move/language/tools/move-cli$ cargo nextest
move/language/tools/move-package$ cargo nextest
```