[borrow v2] Alternative design sketch using a checked reference types
This sketches an alternative design to #835 based on new types `global::Ref<T>` and `global::RefMut<T>`. Those types are considered to be known to the compiler, but some checking is defererred to runtime.

I'm not sure yet whether this design is correct, but I think it points at least in the right direction.