[move-compiler] allow the inlining origin to be embedded in spec marker
This commit extends `UnannotatedExp_::Spec` as following:

```diff
- Spec(SpecId, BTreeMap<Var, Type>),
+ Spec(SpecId, Option<Box<SpecIdent>>, BTreeMap<Var, Type>),
```

Where `SpecIdent` marks the origin of the spec block (before inlining)

```rust
pub struct SpecIdent {
    pub module: ModuleIdent,
    pub function: FunctionName,
    pub id: SpecId,
}
```

This commit also re-assigns `SpecId` for each `UnannotatedExp_::Spec`
inlined to maintain that each inline spec block has a unique ID. This is
needed later in the compilation stage.

This behavior is tested with a `spec_inlining.move` test case added in
the `move_check` folder. Previously, this test case hits an assertion in
the compiler.