[Move-prover] Cherry-pick prover related PRs to aptos-main (#887)
* add support for spec funs

* allow global var to be assigned in spec blocks.

* [move-prover] data invariant assume and assertion for `&mut` params

For functions with `&mut` params `s: S`, the data invariant for `S` will
be assumed at function entry (via `Wellformedness` - this is already
done) and re-checked on any write-back to this reference --- this is
newly added in this commit.

This PR is the opposite of PR #783 which does the opposite, i.e., drop
the assumption at the function entry. PR #783 is proposed following the
following guideline:

> Data invariants are only checked when the `&mut` borrow ends. During
> the lifetime of the `&mut` reference, the invariant can be temporarily
> violated and hence, assuming data invariants there can lead to
> inconsistency issues.

In theory, both PR's fix are better than the current practice (which
is essentially assuming without assertion).

However, as the test result in PR #783 shows, the `Options.move` in
`stdlib` heavily relies on the entrypoint assumption that a `&mut
Option` has either 0 or 1 entry in its vector. Therefore, in this
commit, we tighten the conditions where we assert data invariant.

In summary, after this PR, data invariants will be checked:

- on write-backs to local root, arised from `let r = &mut s` where `s`
  is a local variable
- on write-backs to global root, arised from `let r =
  borrow_global_mut<S>(addr)`
- on write-backs to `&mut` parameter.

```move
struct S {
    f: u8
}
spec S {
    invariant f != 0;
}

fun foo(s: &mut S) {
    // spec: assume s.f != 0;

    let r = &mut s.f;
    r = 0;

    // spec: write-back[s.f](r);
    // spec: assert s.f != 0;
}
```

* fixup! [move-prover] data invariant assume and assertion for `&mut` params

* fixup! fixup! [move-prover] data invariant assume and assertion for `&mut` params

* [move-prover] check data invariants for a &mut param ...

... when and only when the param goes out of scope.

Previously, the approach is to check data invariants
**on every write-back to a &mut param**. This is by definition more
restrictive as this is essentially saying that data invariants are not
allowed to be violated even temporarily in funciton body, which might
not be a practical assumption (as shown in the newly added test case.)

* [move-prover] model type_name::get in stdlib

The majority of the Prover work was firstly introduced when Aptos
introduced the `type_info` Move module. Now that the stdlib has a
similar module `type_name`, we should bridge the effort here.

This PR is still a bit rough as it is using hardcoded module address /
function names to identify the type-reflection schemes. We should come
up with more robust / extensible approach later.

* fixup! [move-prover] model type_name::get in stdlib

* [move-prover] special casing havoc in reach-def annalysis

This commit contains two changes to the reach-def analysis pass:
- if a local variable `x := a` and later we discover that `a := b` too,
  then mark that `x` might alias both `a` and `b`. This is done in the
  intra-procedural abstract interpretation so the def of `x` will be
  updated (joined) on a later stage.
- if a local variable `x := a` and `a` might take the value of both
  original and havoc-ed, then mark that `x` is not uniquely aliased.

Both changes prevents copy propogation from `a` to `x`.

* [move-prover] collect types in write-back chain for mono analysis

In mono analysis, an implicit assumption is that all types used in the
function body will appear either in parameters, locals, return tyupes,
or spec exprs.

While this is true and in a rigourous implemention, should hold, it
might be hit by a cornor case where an imcomplete (and potentially)
inconsistent modeling of borrow natives are in place.

An example is shown in the `dynamic_field::borrow_mut` model:

```move
module sui::dynamic_field {
    struct Field<Name: copy + drop + store, Value: store> has key {
        id: UID,
        name: Name.
        value: Value,
    }

    public fun borrow_mut<Name: copy + drop + store, Value: store>(
        object: &mut UID,
        name: Name,
    ): &mut Value {
        let object_addr = object::uid_to_address(object);
        let hash = hash_type_and_key(object_addr, name);
        let field = borrow_child_object_mut<Field<Name, Value>>(object, hash);
        &mut field.value
    }
    spec borrow_mut {
        pragma opaque;
    }

    public(friend) native fun borrow_child_object_mut<Child: key>(object: &mut UID, id: address): &mut Child;
}
```

With a declaration of in the prover CLI:
```rust
        options
            .prover
            .borrow_natives
            .push("dynamic_field::borrow_child_object_mut".to_string());
        options.backend.borrow_aggregates.push(
            move_prover_boogie_backend::options::BorrowAggregate::new(
                "dynamic_field::borrow_child_object_mut".to_string(),
                "GetDynField".to_string(),
                "UpdateDynField".to_string(),
            ),
        );
```

Now, the `opaque` pragma in `borrow_mut` mark that this funciton is
abstract and when this `borrow_mut` is used, it actually causes an
write-back to the internal `Field` struct which does not show-up in the
function locals (or anything that is collected by mono analysis that has
a type).

Of course in an ideal situation, the `borrow_mut` should be modeled
completely such that the `Field` struct is either eliminated from the
borrow chain or the `Field` should be exposed. However, achiving this
complete modeling takes time and this half-backed solution might need
to stay a bit longer. It is also possible that we do need this facility
(of using `Field` but not exposing `Field` in the Prover). In any case,
it merits a PR as a block.

* add timeout

---------

Co-authored-by: Meng Xu <meng.xu.cs@uwaterloo.ca>