[runtime] No more republishing via session
Remove support for republishing an existing module or doing a
compatibility check in that case in `session` and `loader`.  These
codepaths are never used by Sui, which does not permit republishing at
the same address except in very specific edge cases that are handled
externally, and continuing to support these codepaths will further
complicate support for package upgrades (i.e. republishing) on Sui.

This PR also removes any test that relies on or tests republishing, as
well as support in the sandbox for republishing and/or compatibility checks

Test Plan:

```
move$ cargo nextest
```