[vm errors] Propagate status message 2nd try (#976)
* [vm errors] Propagate status message to adapters

This makes the internal status message of VM errors available in `VMStatus` and accessible by adapters. The purpose is to allow adapters to show more detail of errors like they appear e.g. during bytecode verification. These errors happen more often in practice as anticipated, mainly because uses operate with different versions of networks and tools, e.g. deploying code to some chain where it is not deserializable, or using wrong dependencies on disk which do not match those on the network. Currently, the errors produced by the system are as unspecific as 'LINKER_ERROR'.

* Also add message to ExecutionFailure