[loader] Cache CompiledModules after self-check
...as opposed to after the link check.  This change means that if a
loaded module does not exist at a particular ID, the loader must
verify the linkage on the compiled module before attempting to load
it (whereas previously, it could assume that the existence of the
compiled module in the cache meant that linkage had been checked and
loading could proceed).

This change is motivated by the fact that in the presence of package
upgrades, the linkage for the compiled module will need to be
re-checked for every set of dependencies that the module is being
linked against.

Test Plan:

```
cargo simtest
```