[loader] Limit search for structs/functions in current module with cursor
Rather than assuming that consecutive structs and functions whose
modules match the runtime ID of the module being loaded belong to that
module, use the fact that we're remembering the point at which we
started adding structs and functions in the `CacheCursor` instead.

This is particularly relevant in the face of relinking where multiple
modules might enter the loader with the same runtime IDs.

Test Plan:

Existing tests -- it's not possible to test the effect of loading a
relinked module with a missing type because the verifier will catch it
before we load.